Update README.md
# Mohmmad Kasim's Data Analysis Portfolio

This repository contains my projects and work samples as an **aspiring Data Analyst**. Each project showcases different skills in **SQL**, **Power BI**, and **MS Excel**, demonstrating my ability to clean, analyze, and visualize data for real-world decision-making. 

## Contents

1. **Sales Analysis Dashboard:** A Power BI dashboard that provides actionable insights from historical sales data.
2. **SQL Database Management:** Optimized SQL database to improve query performance and streamline data operations.
3. **Excel Financial Modeling:** A comprehensive financial model in Excel used for revenue forecasting and budgeting.

The purpose of this portfolio is to demonstrate my proficiency in data analytics tools and techniques and highlight the projects I’ve completed that contributed to the success of various clients. I aim to expand my expertise in the data field and continuously work on new and challenging projects.